# api/openapi.yml
openapi: 3.0.3
info:
  title: Developer Utilities MCP Server
  description: |
    A comprehensive, high-performance utility server that consolidates essential development tools into a single service. 
    This server combines cryptography, data manipulation, text processing, identifier generation, time utilities, 
    and network tools to provide AI agents and developers with a one-stop solution for common development tasks.
    
    The server is designed to be secure, stateless, and highly performant, following enterprise-grade standards 
    while remaining accessible for open-source development.
  version: 1.0.0
  contact:
    name: Developer Utilities MCP Server
    url: https://github.com/example/dev-utilities
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
paths:
...
  /api/v1/text/case:
    post:
      tags:
        - Text Processing
      summary: Convert text case
      description: Convert text between different case formats (UPPERCASE, lowercase, camelCase, etc.)
      operationId: convertCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseConvertRequest'
            examples:
              camelCase:
                summary: Convert to camelCase
                value:
                  content: "hello world example"
                  caseType: "camelCase"
              uppercase:
                summary: Convert to UPPERCASE
                value:
                  content: "hello world"
                  caseType: "UPPERCASE"
      responses:
        '200':
          description: Case conversion completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CaseConvertResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/text/info:
    post:
      tags:
        - Text Processing
      summary: Analyze text
      description: Get detailed text analysis including character count, word count, line count, etc.
      operationId: analyzeText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextAnalyzeRequest'
            examples:
              basic:
                summary: Basic text analysis
                value:
                  content: "Hello world!\nThis is a test."
      responses:
        '200':
          description: Text analysis completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TextInfo'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/v1/text/regex:
    post:
      tags:
        - Text Processing
      summary: Test regular expression
      description: Test regular expression pattern against text and return matches
      operationId: testRegex
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegexRequest'
            examples:
              email_pattern:
                summary: Email pattern matching
                value:
                  content: "Contact us at support@example.com or sales@test.org"
                  pattern: "[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}"
                  flags: "g"
      responses:
        '200':
          description: Regex test completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/RegexResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/v1/text/sort:
    post:
      tags:
        - Text Processing
      summary: Sort text lines
      description: Sort text lines alphabetically or numerically in ascending or descending order
      operationId: sortText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextSortRequest'
            examples:
              alphabetical:
                summary: Alphabetical sorting
                value:
                  content: "zebra\napple\nbanana\ncherry"
                  order: "asc"
                  sortType: "alpha"
      responses:
        '200':
          description: Text sorting completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TextSortResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
  /api/v1/data/json/format:
    post:
      tags:
        - Text Processing
      summary: Format JSON
      description: Format or minify JSON with configurable indentation
      operationId: formatJSON
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JSONFormatRequest'
            examples:
              format:
                summary: Format JSON with indentation
                value:
                  content: '{"name":"John","age":30,"city":"New York"}'
                  action: "format"
                  indent: 2
              minify:
                summary: Minify JSON
                value:
                  content: |
                    {
                      "name": "John",
                      "age": 30,
                      "city": "New York"
                    }
                  action: "minify"
      responses:
        '200':
          description: JSON formatting completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/JSONFormatResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

  # Data Transformation Endpoints
...
  /api/v1/web/url:
    post:
      tags:
        - Network Utilities
      summary: Parse or build URLs
      description: Parse URLs into components or build URLs from components
      operationId: processURL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/URLRequest'
            examples:
              parse:
                summary: Parse URL
                value:
                  action: "parse"
                  url: "https://example.com:8080/path?param=value#fragment"
              build:
                summary: Build URL
                value:
                  action: "build"
                  parts:
                    scheme: "https"
                    host: "example.com"
                    path: "/api/v1"
                    query:
                      param1: "value1"
                      param2: "value2"
      responses:
        '200':
          description: URL operation completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        oneOf:
                          - $ref: '#/components/schemas/URLParts'
                          - $ref: '#/components/schemas/URLBuildResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/network/headers:
    post:
      tags:
        - Network Utilities
      summary: Get HTTP headers
      description: Make HTTP GET request and return response headers (with SSRF protection)
      operationId: getHTTPHeaders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeaderRequest'
            examples:
              basic:
                summary: Get headers from URL
                value:
                  url: "https://httpbin.org/headers"
      responses:
        '200':
          description: HTTP headers retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HeaderResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/network/dns:
    post:
      tags:
        - Network Utilities
      summary: DNS lookup
      description: Perform DNS lookup for various record types (A, AAAA, MX, TXT, NS, CNAME)
      operationId: dnsLookup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DNSRequest'
            examples:
              a_record:
                summary: A record lookup
                value:
                  domain: "example.com"
                  recordType: "A"
              mx_record:
                summary: MX record lookup
                value:
                  domain: "example.com"
                  recordType: "MX"
      responses:
        '200':
          description: DNS lookup completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DNSResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/network/ip:
    post:
      tags:
        - Network Utilities
      summary: Analyze IP address
      description: Validate and classify IP addresses (private, public, loopback, version)
      operationId: analyzeIP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IPAnalyzeRequest'
            examples:
              ipv4_public:
                summary: Public IPv4 address
                value:
                  ip: "8.8.8.8"
              ipv4_private:
                summary: Private IPv4 address
                value:
                  ip: "192.168.1.1"
              ipv6:
                summary: IPv6 address
                value:
                  ip: "2001:4860:4860::8888"
      responses:
        '200':
          description: IP analysis completed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IPInfo'
        '400':
          $ref: '#/components/responses/ValidationError'
components:
  schemas:
...
    SuccessResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          description: Response data (varies by endpoint)
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code for programmatic handling
          enum:
            - VALIDATION_ERROR
            - AUTHENTICATION_ERROR
            - AUTHORIZATION_ERROR
            - RATE_LIMIT_ERROR
            - INTERNAL_ERROR
            - SERVICE_UNAVAILABLE
            - NOT_FOUND
            - METHOD_NOT_ALLOWED
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request format"
        details:
          type: string
          description: Additional error details
          example: "Field 'algorithm' is required"

    # Cryptography Schemas
...
    CaseConvertRequest:
      type: object
      required:
        - content
        - caseType
      properties:
        content:
          type: string
          description: Text content to convert
          example: "hello world example"
        caseType:
          type: string
          description: Target case format
          enum: 
            - UPPERCASE
            - lowercase
            - "Title Case"
            - "Sentence case"
            - camelCase
            - PascalCase
            - snake_case
            - kebab-case
          example: "camelCase"
    CaseConvertResponse:
      type: object
      required:
        - result
      properties:
        result:
          type: string
          description: Converted text
          example: "helloWorldExample"
    TextAnalyzeRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          description: Text content to analyze
          example: "Hello world!\nThis is a test."
    TextInfo:
      type: object
      required:
        - characterCount
        - wordCount
        - lineCount
        - sentenceCount
        - byteSize
      properties:
        characterCount:
          type: integer
          description: Number of characters (Unicode-aware)
          example: 25
        wordCount:
          type: integer
          description: Number of words
          example: 5
        lineCount:
          type: integer
          description: Number of lines
          example: 2
        sentenceCount:
          type: integer
          description: Number of sentences
          example: 2
        byteSize:
          type: integer
          description: Size in bytes
          example: 25
    RegexRequest:
      type: object
      required:
        - content
        - pattern
      properties:
        content:
          type: string
          description: Text content to test against
          example: "Contact us at support@example.com or sales@test.org"
        pattern:
          type: string
          description: Regular expression pattern
          example: "[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}"
        flags:
          type: string
          description: Regex flags
          example: "g"
    RegexResponse:
      type: object
      required:
        - matches
        - pattern
      properties:
        matches:
          type: array
          items:
            type: string
          description: Array of matches found
          example: ["support@example.com", "sales@test.org"]
        pattern:
          type: string
          description: Pattern that was used
          example: "[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}"
    TextSortRequest:
      type: object
      required:
        - content
        - order
        - sortType
      properties:
        content:
          type: string
          description: Text content with lines to sort
          example: "zebra\napple\nbanana\ncherry"
        order:
          type: string
          description: Sort order
          enum: [asc, desc]
          example: "asc"
        sortType:
          type: string
          description: Sort type
          enum: [alpha, numeric]
          example: "alpha"
    TextSortResponse:
      type: object
      required:
        - result
      properties:
        result:
          type: string
          description: Sorted text
          example: "apple\nbanana\ncherry\nzebra"
    JSONFormatRequest:
      type: object
      required:
        - content
        - action
      properties:
        content:
          type: string
          description: JSON content to format
          example: '{"name":"John","age":30,"city":"New York"}'
        action:
          type: string
          description: Format action
          enum: [format, minify]
          example: "format"
        indent:
          type: integer
          description: Indentation spaces (for format action)
          minimum: 0
          maximum: 8
          default: 2
          example: 2
    JSONFormatResponse:
      type: object
      required:
        - result
      properties:
        result:
          type: string
          description: Formatted JSON
          example: |
            {
              "name": "John",
              "age": 30,
              "city": "New York"
            }

    # Data Transformation Schemas
...
    URLRequest:
      type: object
      required:
        - action
      properties:
        action:
          type: string
          description: Operation to perform
          enum: [parse, build]
          example: "parse"
        url:
          type: string
          description: URL to parse (required for parse action)
          example: "https://example.com:8080/path?param=value#fragment"
        parts:
          $ref: '#/components/schemas/URLParts'
    URLParts:
      type: object
      required:
        - scheme
        - host
        - path
        - query
        - fragment
      properties:
        scheme:
          type: string
          description: URL scheme
          example: "https"
        host:
          type: string
          description: Hostname and port
          example: "example.com:8080"
        path:
          type: string
          description: URL path
          example: "/path"
        query:
          type: object
          additionalProperties:
            type: string
          description: Query parameters
          example:
            param: "value"
        fragment:
          type: string
          description: URL fragment
          example: "fragment"
    URLBuildResponse:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          description: Built URL
          example: "https://example.com:8080/path?param=value#fragment"
    HeaderRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          description: URL to get headers from
          example: "https://httpbin.org/headers"
    HeaderResponse:
      type: object
      required:
        - url
        - statusCode
        - headers
      properties:
        url:
          type: string
          description: Requested URL
          example: "https://httpbin.org/headers"
        statusCode:
          type: integer
          description: HTTP status code
          example: 200
        headers:
          type: object
          additionalProperties:
            type: string
          description: Response headers
          example:
            Content-Type: "application/json"
            Server: "nginx/1.18.0"
    DNSRequest:
      type: object
      required:
        - domain
        - recordType
      properties:
        domain:
          type: string
          description: Domain name to lookup
          example: "example.com"
        recordType:
          type: string
          description: DNS record type
          enum: [A, AAAA, MX, TXT, NS, CNAME]
          example: "A"
    DNSResponse:
      type: object
      required:
        - domain
        - recordType
        - records
      properties:
        domain:
          type: string
          description: Domain that was queried
          example: "example.com"
        recordType:
          type: string
          description: Record type that was queried
          example: "A"
        records:
          type: array
          items:
            type: string
          description: DNS records found
          example: ["93.184.216.34"]
        ttl:
          type: integer
          description: Time to live in seconds
          example: 3600
    IPAnalyzeRequest:
      type: object
      required:
        - ip
      properties:
        ip:
          type: string
          description: IP address to analyze
          example: "8.8.8.8"
    IPInfo:
      type: object
      required:
        - ip
        - version
        - isPrivate
        - isPublic
        - isLoopback
      properties:
        ip:
          type: string
          description: IP address that was analyzed
          example: "8.8.8.8"
        version:
          type: integer
          description: IP version (4 or 6)
          example: 4
        isPrivate:
          type: boolean
          description: Whether IP is in private range
          example: false
        isPublic:
          type: boolean
          description: Whether IP is public
          example: true
        isLoopback:
          type: boolean
          description: Whether IP is loopback
          example: false
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missing_field:
              summary: Missing required field
              value:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Invalid request format"
                  details: "Field 'algorithm' is required"
            invalid_value:
              summary: Invalid field value
              value:
                success: false
                error:
                  code: "VALIDATION_ERROR"
                  message: "Invalid request format"
                  details: "Field 'algorithm' must be one of: md5, sha1, sha256, sha512"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            generic:
              summary: Generic internal error
              value:
                success: false
                error:
                  code: "INTERNAL_ERROR"
                  message: "Internal server error"
                  details: "An unexpected error occurred"
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication